nxp-imx8: upgrade the script used to generate graphic layer
Issue: LINUXEXEC-6752

Because the license limitation of the nxp graphic library, it is not
allowed to integrate into wrlinux product. So, create a script to
generate the graphic layer on customer's side and notice them
to pay attention to the license limitation.
In wrlinux1019, upgrade this script based on NXP-SDK5.4.

Signed-off-by: Meng Li <Meng.Li@windriver.com>