nxp-imx8: put the dtb file to higher address
Issue: LIN1019-4515

When build this BSP with kts layer, there are more kernel configs
enabled. So, kernel become larger, and it will use about 46M region.
But with the previous boot parameter, the space region between
kernel loading address and dtb loading address is about 43M, so it
is not enough space for current kernel image that is build with kts layer.
Therefore, put the dtb file at higher address so that extend the space
region kernel loading address and dtb loading address.

Signed-off-by: Meng Li <Meng.Li@windriver.com>