nxp-imx8: put the dtb file to higher address for imx8mini platform
Issue: LIN1019-5008

When build this BSP with kts layer, there are more kernel configs
enabled. So, kernel becomes larger, use the previous boot parameter,
it is not enough space for current kernel image that is built with kts
layer. Therefore, put the dtb file at higher address so that extend
the space region for loading kernel.

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>